Update vaccination data
Note that the 2021-01-01 data is copied from 2020-12-31 because
there was no report published on 2021-01-02. The RKI announced
that they will only update the data on workdays. To handle this,
we’ll fill the gaps by copying data from the previous published
day. Due to the 7-day moving average of the derivative we use
in the calculations, this’ll be OK.