Reduce memory footprint of rki-to-influx by a factor of 10
This is because we now don’t hold the full cartesian product of
states and districts in memory; instead, we treat the
state+district pair as one value. This also allows us to remove
the ugly check in generate_counter_samples as we now have much
more normalised data.

The theoretical improvement would be a bit less than a factor of
15 (16 states, i.e. the data is kept in memory 15 times more than
needed).