Fold the different Berlin districts into one
The separation is a thing created artificially by the RKI and not a
thing any other german authority uses for any kind of statistics. The
problem is that once we mix separated and non-separated data we get into
trouble when using sum() functions in the dashboard.

To avoid that, we go the only way we can: aggregate Berlin.