Add support for a correct cumsum for pub series
Previously, the cumsum had way too low values for those series generated
from the publication date. The reason for that is that we generally do
not have all data available, so we can only start the publication
date based time series at a certain point (currently 2020-12-19).

This is problematic when comparing the cumsum of the publication date
series with other cumsums or the population. In order to fix this, we
adapt the generation process to include a "base file" which is generated
from the report data of the first date from which the publication date
time series is created. That way, the cumsum from before the first
publication date report is identical to the cumsum of the report date
series, and only after publication date information is available, the
cumsum differs.

This allows for some nice things.